backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==41.0.7
defusedxml==0.7.1
Django==3.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
mccabe==0.7.0
oauthlib==3.2.2
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.1
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
